добавлено задание №15 Изменено задание №6

+++ src/arithmetic_operations.java
@@ -6,12 +6,12 @@ public class arithmetic_operations {
 
         Scanner in = new Scanner(System.in);
         System.out.println("Введите число x: ");
-        int num_x = in.nextInt();
+        double num_x = in.nextInt();
         System.out.println("Введите число y: ");
-        int num_y = in.nextInt();
+        double num_y = in.nextInt();
         System.out.println("Введите число z: ");
-        int num_z = in.nextInt();
-        int averageNum = (num_x + num_y + num_z)/3;
+        double num_z = in.nextInt();
+        double averageNum = (num_x + num_y + num_z)/3;
         System.out.println("Среднее арифметическое трех чисел x,y,z = "+ averageNum);
         if (averageNum/2 > 3){
             System.out.println("Программа выполнена корректно");
